manage file changes to meet lines changed requirement
